requirements.txt
fastapi==0.101.1
PyMySQL==1.1.0
SQLAlchemy==2.0.20
passlib==1.7.4
bcrypt==4.0.1
pydantic-settings==2.0.3
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
uvicorn==0.23.2
alembic==1.11.3
cryptography=41.0.3
psycopg2-binary
FastAPI-SQLAlchemy=0.2.1
pydantic